Storybook config and files updated to include create-react-app addon and removing preset-scss
