+++ .storybook/main.js
@@ -1,10 +1,11 @@
 module.exports = {
   "stories": [
-    "../src/components/**/*.stories.@(js|jsx|ts|tsx)"
+    "../src/components/**/*.stories.@(jsx)"
   ],
   "addons": [
     "@storybook/addon-links",
     "@storybook/addon-essentials",
-    "@storybook/preset-scss"
-  ]
+    // "@storybook/preset-scss",
+    "@storybook/preset-create-react-app"
+  ],
 }
