+++ package.json
@@ -47,7 +47,6 @@
     "@storybook/addon-essentials": "^6.3.12",
     "@storybook/addon-links": "^6.3.12",
     "@storybook/preset-create-react-app": "^3.2.0",
-    "@storybook/preset-scss": "^1.0.3",
     "@storybook/react": "^6.3.12",
     "babel-loader": "8.1.0",
     "eslint": "^7.15.0",
