+++ src/components/atoms/basic-button/index.stories.jsx
@@ -1,3 +1,4 @@
+import React from 'react'
 import BasicButton from '.'
 // APP_ROUTES
 import { STORYBOOK_ROUTES } from '../../../constants/routes.json'
