+++ src/components/atoms/message-block/index.stories.jsx
@@ -1,3 +1,4 @@
+import React from 'react'
 import MessageBlock from '.'
 // APP_ROUTES
 import { STORYBOOK_ROUTES } from '../../../constants/routes.json'
