+++ src/components/atoms/title/index.stories.jsx
@@ -1,3 +1,4 @@
+import React from 'react'
 import Title from '.'
 // APP_ROUTES
 import { STORYBOOK_ROUTES } from '../../../constants/routes.json'
