+++ src/components/organisms/basic-frame/index.stories.jsx
@@ -1,3 +1,4 @@
+import React from 'react'
 import BasicFrame from '.'
 import ButtonGroup from '../../molecules/button-group'
 // APP_ROUTES
